refactor: StudentInfoResponse fix
