+++ src/main/java/com/tenbit/beep/classroom/service/impl/ClassRoomServiceImpl.java
@@ -2,7 +2,7 @@ package com.tenbit.beep.classroom.service.impl;
 
 import com.tenbit.beep.auth.domain.User;
 import com.tenbit.beep.classroom.dto.LookUpStudentsByClassRequest;
-import com.tenbit.beep.common.dto.StudentInfoResponse;
+import com.tenbit.beep.classroom.dto.StudentInfoResponse;
 import com.tenbit.beep.classroom.repository.ClassRoomRepository;
 import com.tenbit.beep.classroom.service.ClassRoomService;
 import com.tenbit.beep.common.exception.UserNotFoundException;
