+++ src/main/java/com/tenbit/beep/clubroom/service/impl/ClubRoomServiceImpl.java
@@ -1,14 +1,12 @@
 package com.tenbit.beep.clubroom.service.impl;
 
 import com.tenbit.beep.auth.domain.User;
-import com.tenbit.beep.clubroom.domain.ClubRoom;
 import com.tenbit.beep.clubroom.dto.LookUpStudentsByClubRequest;
 import com.tenbit.beep.clubroom.repository.ClubRoomRepository;
 import com.tenbit.beep.clubroom.service.ClubRoomService;
-import com.tenbit.beep.common.dto.StudentInfoResponse;
+import com.tenbit.beep.classroom.dto.StudentInfoResponse;
 import com.tenbit.beep.common.exception.UserNotFoundException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
